Selva Oscura: CI clang-tidy fixes - rename BehaviorTree/EnemyArchetype members to drop trailing underscore (readability-identifier-naming), NOLINT va_list false positives in PoseSampler+CombatLog, cast %zu format arg to size_t for LP64 Linux, fix NOLINTNEXTLINE placement (multi-line comments were splitting suppression from target line), disable clang-analyzer-unix.Malloc globally (cgltf's free pattern not modeled)

+++ games/selva-oscura/src/gameplay/PerFrameTick.cpp
@@ -897,7 +897,8 @@ static void writeFrameCaptureContactSheet(int n_frames, const std::string& dir)
         std::fprintf(stderr,
                      "[frame-capture] sheet would be %zu MB (%dx%d %dx%d cells); "
                      "skipping. Individual PNGs at %s/\n",
-                     bytes / (1024ULL * 1024ULL), sheet_w, sheet_h, cols, rows, dir.c_str());
+                     static_cast<std::size_t>(bytes / (1024ULL * 1024ULL)), sheet_w, sheet_h, cols,
+                     rows, dir.c_str());
         return;
     }
     std::vector<unsigned char> sheet(bytes, 32);
